// llm-graph-builder/frontend/src/types.ts
export type Scheme = Record<string, string>;
// ...
export type BasicRelationship = {
  id: string;
  to: string;
  from: string;
  type: string;
  caption: string;
};

export type BasicNode = {
  id: string;
  type: string;
  labels: string[];
  properties: Record<string, string>;
  propertyTypes: Record<string, string>;
};

export type PropertyValue = string | number | boolean | [];

export type GraphPropertiesTableProps = {
  propertiesWithTypes: {
    key: string;
    value: PropertyValue;
  }[];
};

export type GraphPropertiesPanelProps = {
  inspectedItem: BasicNode | BasicRelationship;
  newScheme: Scheme;
};